+++ engine/renderercommon/tr_image_png.cpp
@@ -319,7 +319,7 @@ static void *BufferedFileRead(struct BufferedFile *BF, unsigned Length)
 	 *  not enough bytes left
 	 */
 
-	if(Length > BF->BytesLeft)
+	if(Length > (unsigned int)BF->BytesLeft)
 	{
 		return(NULL);
 	}
@@ -413,7 +413,7 @@ static qboolean BufferedFileSkip(struct BufferedFile *BF, unsigned Offset)
 	 *  We can only skip to the end of the BufferedFile.
 	 */
 
-	if(Offset > BF->BytesLeft)
+	if(Offset > (unsigned int)BF->BytesLeft)
 	{
 		return(qfalse);
 	}
