+++ engine/renderercommon/tr_image_tga.cpp
@@ -148,7 +148,7 @@ void R_LoadTGA ( const char *name, byte **pic, int *width, int *height)
 			case 8:
 				for ( row = rows - 1; row >= 0; row-- )	{
 					pixbuf = targa_rgba + row * columns * 4;
-					for ( column = 0; column < columns; column++ ) {
+					for ( column = 0; (unsigned int)column < columns; column++ ) {
 						byte red, green, blue;
 						red = green = blue = *buf_p++;
 						*pixbuf++ = red;
@@ -161,7 +161,7 @@ void R_LoadTGA ( const char *name, byte **pic, int *width, int *height)
 			case 24:
 				for ( row = rows - 1; row >= 0; row-- ) {
 					pixbuf = targa_rgba + row * columns * 4;
-					for ( column = 0; column < columns; column++ ) {
+					for ( column = 0; (unsigned int)column < columns; column++ ) {
 						byte red, green, blue;
 						blue = *buf_p++;
 						green = *buf_p++;
@@ -176,7 +176,7 @@ void R_LoadTGA ( const char *name, byte **pic, int *width, int *height)
 			case 32:
 				for ( row = rows - 1; row >= 0; row-- ) {
 					pixbuf = targa_rgba + row * columns * 4;
-					for ( column = 0; column < columns; column++ ) {
+					for ( column = 0; (unsigned int)column < columns; column++ ) {
 						byte red, green, blue, alpha;
 						blue = *buf_p++;
 						green = *buf_p++;
@@ -199,7 +199,7 @@ void R_LoadTGA ( const char *name, byte **pic, int *width, int *height)
 
 		for(row=rows-1; row>=0; row--) {
 			pixbuf = targa_rgba + row*columns*4;
-			for(column=0; column<columns; ) {
+			for(column=0; (unsigned int)column<columns; ) {
 				if(buf_p + 1 > end)
 					ri.Error (ERR_DROP, "LoadTGA: file truncated (%s)", name);
 				packetHeader= *buf_p++;
