+++ engine/server/sv_main.cpp
@@ -83,7 +83,7 @@ static const char *SV_ExpandNewlines( const char *in ) {
 	int		l;
 
 	l = 0;
-	while ( *in && l < sizeof(string) - 3 ) {
+	while ( *in && (size_t)l < sizeof(string) - 3 ) {
 		if ( *in == '\n' ) {
 			string[l++] = '\\';
 			string[l++] = 'n';
@@ -467,7 +467,7 @@ static leakyBucket_t *SVC_BucketForAddress( const netadr_t *address, int burst [
 		interval = now - bucket->rate.lastTime;
 
 		// Reclaim expired buckets
-		if ( bucket->type != NA_BAD && (unsigned)interval > ( bucket->rate.burst * period ) ) {
+		if ( bucket->type != NA_BAD && (unsigned)interval > (unsigned int)( ( bucket->rate.burst * period ) ) ) {
 			if ( bucket->prev != NULL ) {
 				bucket->prev->next = bucket->next;
 			} else {
